chore: shuffle around classes in the public api to reduce breakages, make unleash builder work for proxy sdk and collapse a few interfaces
