+++ examples/09-cached-proxy.php
@@ -2,7 +2,7 @@
 
 require __DIR__ . '/_common.php';
 
-use Unleash\Client\ProxyUnleashBuilder;
+use Unleash\Client\UnleashBuilder;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\Cache\Psr16Cache;
 
@@ -10,7 +10,7 @@ use Symfony\Component\Cache\Psr16Cache;
 $psr6Cache = new FilesystemAdapter();
 $cache = new Psr16Cache($psr6Cache);
 
-$unleash = ProxyUnleashBuilder::create()
+$unleash = UnleashBuilder::create()
     ->withAppName($appName)
     ->withAppUrl($appUrl)
     ->withInstanceId($instanceId)
@@ -19,6 +19,7 @@ $unleash = ProxyUnleashBuilder::create()
     ->withCacheTimeToLive(3)
     ->withMetricsInterval(300)
     ->withHeader('Authorization', $apiKey)
+    ->withProxy($apiKey)
     ->build();
 
 if ($unleash->isEnabled('myFeature')) {
