+++ src/DTO/DefaultResolvedVariant.php
@@ -3,22 +3,13 @@
 namespace Unleash\Client\DTO;
 
 use JetBrains\PhpStorm\ArrayShape;
-use JetBrains\PhpStorm\ExpectedValues;
-use Unleash\Client\Enum\Stickiness;
 
-final class DefaultInternalVariant implements InternalVariant
+final class DefaultResolvedVariant implements ResolvedVariant
 {
-    /**
-     * @param array<VariantOverride> $overrides
-     */
     public function __construct(
         private readonly string $name,
         private readonly bool $enabled,
-        private readonly int $weight = 0,
-        #[ExpectedValues(valuesFromClass: Stickiness::class)]
-        private readonly string $stickiness = Stickiness::DEFAULT,
         private readonly ?VariantPayload $payload = null,
-        private readonly ?array $overrides = null,
     ) {
     }
 
@@ -50,27 +41,8 @@ final class DefaultInternalVariant implements InternalVariant
         return $result;
     }
 
-    public function getWeight(): int
-    {
-        return $this->weight;
-    }
-
     public function isEnabled(): bool
     {
         return $this->enabled;
-    }
-
-    /**
-     * @return array<VariantOverride>
-     */
-    public function getOverrides(): array
-    {
-        return $this->overrides ?? [];
-    }
-
-    #[ExpectedValues(valuesFromClass: Stickiness::class)]
-    public function getStickiness(): string
-    {
-        return $this->stickiness;
     }
 }
