+++ src/DTO/DefaultVariant.php
@@ -3,13 +3,22 @@
 namespace Unleash\Client\DTO;
 
 use JetBrains\PhpStorm\ArrayShape;
+use JetBrains\PhpStorm\ExpectedValues;
+use Unleash\Client\Enum\Stickiness;
 
 final class DefaultVariant implements Variant
 {
+    /**
+     * @param array<VariantOverride> $overrides
+     */
     public function __construct(
         private readonly string $name,
         private readonly bool $enabled,
+        private readonly int $weight = 0,
+        #[ExpectedValues(valuesFromClass: Stickiness::class)]
+        private readonly string $stickiness = Stickiness::DEFAULT,
         private readonly ?VariantPayload $payload = null,
+        private readonly ?array $overrides = null,
     ) {
     }
 
@@ -41,8 +50,27 @@ final class DefaultVariant implements Variant
         return $result;
     }
 
+    public function getWeight(): int
+    {
+        return $this->weight;
+    }
+
     public function isEnabled(): bool
     {
         return $this->enabled;
+    }
+
+    /**
+     * @return array<VariantOverride>
+     */
+    public function getOverrides(): array
+    {
+        return $this->overrides ?? [];
+    }
+
+    #[ExpectedValues(valuesFromClass: Stickiness::class)]
+    public function getStickiness(): string
+    {
+        return $this->stickiness;
     }
 }
