--- src/DTO/ProxyFeature.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace Unleash\Client\DTO;
-
-interface ProxyFeature
-{
-    public function getName(): string;
-
-    public function isEnabled(): bool;
-
-    public function getVariant(): Variant;
-
-    public function hasImpressionData(): bool;
-}
