+++ src/DTO/Variant.php
@@ -2,7 +2,9 @@
 
 namespace Unleash\Client\DTO;
 
+use JetBrains\PhpStorm\ExpectedValues;
 use JsonSerializable;
+use Unleash\Client\Enum\Stickiness;
 
 interface Variant extends JsonSerializable
 {
@@ -11,4 +13,14 @@ interface Variant extends JsonSerializable
     public function isEnabled(): bool;
 
     public function getPayload(): ?VariantPayload;
+
+    public function getWeight(): int;
+
+    /**
+     * @return array<VariantOverride>
+     */
+    public function getOverrides(): array;
+
+    #[ExpectedValues(valuesFromClass: Stickiness::class)]
+    public function getStickiness(): string;
 }
