+++ src/Repository/DefaultUnleashProxyRepository.php
@@ -1,0 +1,218 @@
+<?php
+
+namespace Unleash\Client\Repository;
+
+use JsonException;
+use Psr\Http\Client\ClientExceptionInterface;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestFactoryInterface;
+use Psr\SimpleCache\InvalidArgumentException;
+use Unleash\Client\Configuration\Context;
+use Unleash\Client\Configuration\UnleashConfiguration;
+use Unleash\Client\Configuration\UnleashContext;
+use Unleash\Client\DTO\DefaultProxyFeature;
+use Unleash\Client\DTO\Feature;
+
+final class DefaultUnleashProxyRepository implements UnleashRepository, ProxyRepository
+{
+    public function __construct(
+        private readonly DefaultUnleashRepository $baseRepo,
+        private UnleashConfiguration $configuration,
+        private ClientInterface $httpClient,
+        private RequestFactoryInterface $requestFactory
+    ) {
+    }
+
+    public function findFeature(string $featureName): ?Feature
+    {
+        return $this->baseRepo->findFeature($featureName);
+    }
+
+    /**
+     * @throws InvalidArgumentException
+     * @throws ClientExceptionInterface
+     * @throws JsonException
+     *
+     * @return iterable<Feature>
+     */
+    public function getFeatures(): iterable
+    {
+        return $this->baseRepo->getFeatures();
+    }
+
+    public function findFeatureByContext(string $featureName, ?Context $context = null): ?Feature
+    {
+        $cacheKey = $this->getCacheKey($featureName, $context);
+
+        if ($this->configuration->getCache() !== null && $this->configuration->getCache()->has($cacheKey)) {
+            $cachedFeature = $this->configuration->getCache()->get($cacheKey);
+            if (is_array($cachedFeature)) {
+                $featureData = $this->validateResponse($cachedFeature);
+                if ($featureData !== null) {
+                    return new DefaultProxyFeature($featureData);
+                }
+            }
+        }
+
+        $context ??= new UnleashContext();
+        $featureUrl = $this->configuration->getUrl() . 'frontend/features/' . $featureName;
+        $url = $this->addQuery($featureUrl, $this->contextToQueryString($context));
+
+        $request = $this->requestFactory->createRequest('GET', $url)
+            ->withHeader('Content-Type', 'application/json')
+            ->withHeader('Accept', 'application/json');
+
+        foreach ($this->configuration->getHeaders() as $name => $value) {
+            $request = $request->withHeader($name, $value);
+        }
+
+        $request = $request->withHeader('Authorization', $this->configuration->getProxyKey());
+
+        $response = $this->httpClient->sendRequest($request);
+        $body = json_decode($response->getBody(), true);
+        if ($body === null) {
+            return null;
+        }
+        if (is_array($body)) {
+            $featureData = $this->validateResponse($body);
+
+            if ($featureData !== null) {
+                if ($this->configuration->getCache() !== null) {
+                    $this->configuration->getCache()->set($cacheKey, $featureData);
+                }
+
+                return new DefaultProxyFeature($featureData);
+            }
+        }
+
+        return null;
+    }
+
+    private function getCacheKey(string $featureName, ?Context $context): string
+    {
+        if($context === null) {
+            return $featureName;
+        }
+        $contextHash = md5($this->contextToQueryString($context));
+
+        return $featureName . '-' . $contextHash;
+    }
+
+    private function contextToQueryString(Context $context): string
+    {
+        $query = [];
+        $values = [
+            'userId' => $context->getCurrentUserId(),
+            'sessionId' => $context->getSessionId(),
+            'remoteAddress' => $context->getIpAddress(),
+        ];
+
+        $values = array_filter($values, fn (?string $value) => $value !== null);
+        $properties = array_filter($context->getCustomProperties(), fn (?string $value) => $value !== null);
+
+        foreach ($values as $key => $value) {
+            $query[$key] = $value;
+        }
+
+        $query['properties'] = [];
+
+        foreach ($properties as $key => $value) {
+            $query['properties'][$key] = $value;
+        }
+
+        return urldecode(http_build_query($query));
+    }
+
+    private function addQuery(string $url, string $query): string
+    {
+        $urlParts = parse_url($url);
+
+        // The base url is already validated, so this should only happen if a feature toggle has a name
+        // that can't form part of a URL but this shouldn't ever happen, given that this is validated
+        // by the Unleash server and UI
+
+        // @codeCoverageIgnoreStart
+        if ($urlParts === false) {
+            throw new \InvalidArgumentException('Invalid URL provided');
+        }
+        // @codeCoverageIgnoreEnd
+
+        if (!isset($urlParts['query'])) {
+            $urlParts['query'] = $query;
+        }
+
+        return $this->buildUrl($urlParts);
+    }
+
+    /**
+     * @param array<string, mixed> $response
+     *
+     * @return array{
+     *     name: string,
+     *     enabled: bool,
+     *     variant: array{
+     *         name: string,
+     *         enabled: bool,
+     *         payload?: array{
+     *             type: string,
+     *             value: string
+     *         }
+     *     },
+     *     impression_data: bool
+     * }|null
+     */
+    private function validateResponse(array $response): ?array
+    {
+        if (!isset($response['name'], $response['enabled'], $response['variant'], $response['impression_data'])) {
+            return null;
+        }
+
+        if (!is_string($response['name']) || !is_bool($response['enabled']) || !is_bool($response['impression_data']) || !is_array($response['variant'])) {
+            return null;
+        }
+
+        if (!isset($response['variant']['name'], $response['variant']['enabled']) || !is_string($response['variant']['name']) || !is_bool($response['variant']['enabled'])) {
+            return null;
+        }
+
+        if (isset($response['variant']['payload']) && (!is_array($response['variant']['payload']) || !isset($response['variant']['payload']['type'], $response['variant']['payload']['value']) || !is_string($response['variant']['payload']['type']) || !is_string($response['variant']['payload']['value']))) {
+            return null;
+        }
+
+        return $response;
+    }
+
+    /**
+     * @param array<string, mixed> $parts
+     *
+     * @return string
+     */
+    private function buildUrl(array $parts): string
+    {
+        $result = '';
+        if (isset($parts['scheme']) && is_string($parts['scheme'])) {
+            $result .= $parts['scheme'] . '://';
+        }
+        if (isset($parts['user']) && is_string($parts['user'])) {
+            $result .= $parts['user'];
+            if (isset($parts['pass']) && is_string($parts['pass'])) {
+                $result .= ':' . $parts['pass'];
+            }
+            $result .= '@';
+        }
+        if (isset($parts['host']) && is_string($parts['host'])) {
+            $result .= $parts['host'];
+        }
+        if (isset($parts['port']) && is_numeric($parts['port'])) {
+            $result .= ':' . $parts['port'];
+        }
+        if (isset($parts['path']) && is_string($parts['path'])) {
+            $result .= $parts['path'];
+        }
+        if (isset($parts['query']) && is_string($parts['query'])) {
+            $result .= '?' . $parts['query'];
+        }
+
+        return $result;
+    }
+}
